Congestion alleviating approach to claiming contract. (#11)
Added an example of how we can do zora token claiming for the community using a mapping of account -> allocation instread of a merkle tree. 

The main difference is wed need to batch insert allocations as arrays instead of inserting a merkle tree.

The advantage during claiming is that users don't need to submit an entire merkle proof when claiming, which fills up the calldata and blockspace leading to congestion.

We take advantage of the fact that this contract is deployed on base L2, where storage costs are low.  

We've estimated this would take around 1.5eth to fill this data for 2m records on the base network.